Refactor (bootstrap-pull-request)
